complete css flex exercise 5
Note in this exercise I had to pay close attention to containers/their
contents to know which items would be affected by the styling. Each
element is not inherently it's own "child" within the container, it
must be made that way through looking at divs. A child can be the
container to further elements.